Fix protocol consolidation error on tests
Protocol consolidation needs to be turned off when we implement
protocols in the test files.

It's because the consolidation occurs when the project is compiled. The
tests are compiled later (when they need to be run) so the consolidation
is no longer possible.

Anyway, here I just turned the consolidation off for the test environment.

More information here: https://github.com/elixir-lang/elixir/issues/4175